+++ python/ferrmion/__init__.py
@@ -5,7 +5,7 @@ rust functions before importing functions from the python module.
 """
 
 from .core import hartree_fock_state, symplectic_product
-from .encode.base import FermionQubitEncoding
+from .encode import FermionQubitEncoding
 from .encode.knto import KNTO, knto_symplectic_matrix
 from .encode.ternary_tree import TernaryTree
 from .encode.ternary_tree_node import TTNode, node_sorter